+++ lib/app/modules/home/sub_views/view_accounts/view_accounts.dart
@@ -128,6 +128,7 @@ class ViewAccountsState extends ViewForMoneyObjectsState {
                   ),
                 ]),
               ),
+              InternalViewSwitching.toInvestments(accountName: account.name.value),
             ],
           ),
         );