+++ lib/app/modules/home/sub_views/view_accounts/view_accounts.dart
@@ -124,7 +124,7 @@ class ViewAccountsState extends ViewForMoneyObjectsState {
                 filters: FieldFilters([
                   FieldFilter(
                     fieldName: Constants.viewTransactionFieldnameAccount,
-                    filterTextInLowerCase: account.name.value.toLowerCase(),
+                    strings: [account.name.value],
                   ),
                 ]),
               ),